+++ SqExpress.CodegenUtil/UtilArguments.cs
@@ -1,6 +1,0 @@
-﻿using CommandLine;
-
-namespace SqExpress.CodeGenUtil
-{
-
-}