+++ SqExpress.GetStarted/Models/Extra/IModel.cs
@@ -1,0 +1,7 @@
+﻿namespace SqExpress.GetStarted.Models.Extra
+{
+    public interface IModel
+    {
+        
+    }
+}